+++ DEPRECATED/OLDoptimizer_comparisons.py
