+++ output/solver_log.txt
@@ -2,24 +2,24 @@ Welcome to the CBC MILP Solver
 Version: 2.10.3 
 Build Date: Dec 15 2019 
 
-command line - /Users/glennharless/miniconda3/envs/psych_opt/lib/python3.12/site-packages/pulp/solverdir/cbc/osx/64/cbc /var/folders/7m/njx2trb92gs75_vg_zs20px00000gn/T/c3de40f92f194b2c9daeae4dd053642f-pulp.mps -timeMode elapsed -branch -printingOptions all -solution /var/folders/7m/njx2trb92gs75_vg_zs20px00000gn/T/c3de40f92f194b2c9daeae4dd053642f-pulp.sol (default strategy 1)
+command line - /Users/glennharless/miniconda3/envs/psych_opt/lib/python3.12/site-packages/pulp/solverdir/cbc/osx/64/cbc /var/folders/7m/njx2trb92gs75_vg_zs20px00000gn/T/a14e8f6730ac4e0c8af0b49e9b53b1a4-pulp.mps -timeMode elapsed -branch -printingOptions all -solution /var/folders/7m/njx2trb92gs75_vg_zs20px00000gn/T/a14e8f6730ac4e0c8af0b49e9b53b1a4-pulp.sol (default strategy 1)
 At line 2 NAME          MODEL
 At line 3 ROWS
-At line 3358 COLUMNS
-At line 14262 RHS
-At line 17616 BOUNDS
-At line 19296 ENDATA
-Problem MODEL has 3353 rows, 1679 columns and 7544 elements
+At line 1898 COLUMNS
+At line 8057 RHS
+At line 9951 BOUNDS
+At line 10901 ENDATA
+Problem MODEL has 1893 rows, 949 columns and 4259 elements
 Coin0008I MODEL read with 0 errors
 Option for timeMode changed from cpu to elapsed
-Continuous objective value is 19 - 0.01 seconds
+Continuous objective value is 19 - 0.00 seconds
 Cgl0003I 0 fixed, 13 tightened bounds, 0 strengthened rows, 0 substitutions
-Cgl0004I processed model has 436 rows, 292 columns (292 integer (0 of which binary)) and 1162 elements
+Cgl0004I processed model has 358 rows, 240 columns (240 integer (0 of which binary)) and 954 elements
 Cutoff increment increased from 1e-05 to 0.9999
-Cbc0045I 1 integer variables out of 292 objects (292 integer) have cost of 1 - high priority
+Cbc0045I 1 integer variables out of 240 objects (240 integer) have cost of 1 - high priority
 Cbc0045I branch on satisfied Y create fake objective Y random cost Y
-Cbc0012I Integer solution of 20 found by DiveCoefficient after 0 iterations and 0 nodes (0.03 seconds)
-Cbc0001I Search completed - best objective 20, took 0 iterations and 0 nodes (0.03 seconds)
+Cbc0012I Integer solution of 20 found by DiveCoefficient after 0 iterations and 0 nodes (0.01 seconds)
+Cbc0001I Search completed - best objective 20, took 0 iterations and 0 nodes (0.01 seconds)
 Cbc0035I Maximum depth 0, 0 variables fixed on reduced cost
 Cuts at root node changed objective from 20 to 20
 Probing was tried 0 times and created 0 cuts of which 0 were active after adding rounds of cuts (0.000 seconds)
@@ -36,9 +36,9 @@ Result - Optimal solution found
 Objective value:                20.00000000
 Enumerated nodes:               0
 Total iterations:               0
-Time (CPU seconds):             0.03
-Time (Wallclock seconds):       0.04
+Time (CPU seconds):             0.02
+Time (Wallclock seconds):       0.02
 
 Option for printingOptions changed from normal to all
-Total time (CPU seconds):       0.04   (Wallclock seconds):       0.04
+Total time (CPU seconds):       0.02   (Wallclock seconds):       0.02
