--- output/ward_optimization_results.csv
@@ -1,15 +1,0 @@
-Double Rooms,Single Rooms,Wasted Beds
-0,26,118.0
-1,24,314.0
-2,22,562.0
-3,20,894.0
-4,18,1274.0
-5,16,1770.0
-6,14,2344.0
-7,12,3064.0
-8,10,3870.0
-9,8,4812.0
-10,6,5989.0
-11,4,7495.0
-12,2,9141.0
-13,0,10814.0
